Fix multiple audio dirs in wants & Hide dirs
Obfuscate the audio source dir by saving only its hash in the db, and
subsequently the wants file. This is done in order to prevent leaking
information about the server file structure outside of the audio dir.
How vulnerable is this to a pre-image attack? Maybe salt the whole
thing in the future...