Make CoverArt finder caching
Before the CoverArt finder would have searched for a cover image and
created a new instance on every constructor call.

Now the CoverArt class has a "least-recently-used" (LRU) behavior that
returns a previous instance if the image file path is the same. Images
are evicted from the LRU after a configurable amount of other images
are loaded.